Move sync lookup trait function to its caller
